move spawners into configurables dir
